requirements.txt
Django==2.2.1
django-solo==1.1.3
Pillow==6.0.0
pytz==2019.1
sqlparse==0.3.0
